chore(structure): moves current federation setup logic to legacy pkg
As we are introducing `controller-runtime`-based solution we will need
to aggresively refactor and remove some chunks of code. For this reason
moving current solution to dedicated `legacy` directory can help with
identifying what needs to be removed.
