+++ internal/pkg/istio/config_factory.go
@@ -33,7 +33,7 @@ import (
 
 	"github.com/openshift-service-mesh/federation/internal/api/federation/v1alpha1"
 	"github.com/openshift-service-mesh/federation/internal/pkg/config"
-	"github.com/openshift-service-mesh/federation/internal/pkg/fds"
+	"github.com/openshift-service-mesh/federation/internal/pkg/legacy/fds"
 	"github.com/openshift-service-mesh/federation/internal/pkg/networking"
 )
