+++ internal/pkg/legacy/fds/exported_service_generator_test.go
@@ -27,7 +27,7 @@ import (
 
 	"github.com/openshift-service-mesh/federation/internal/api/federation/v1alpha1"
 	"github.com/openshift-service-mesh/federation/internal/pkg/config"
-	"github.com/openshift-service-mesh/federation/internal/pkg/informer"
+	"github.com/openshift-service-mesh/federation/internal/pkg/legacy/informer"
 )
 
 var (
