+++ internal/pkg/legacy/kube/reconciler_manager.go
@@ -20,7 +20,7 @@ import (
 
 	istiolog "istio.io/istio/pkg/log"
 
-	"github.com/openshift-service-mesh/federation/internal/pkg/xds"
+	"github.com/openshift-service-mesh/federation/internal/pkg/legacy/xds"
 )
 
 var log = istiolog.RegisterScope("kube", "Kubernetes reconciler")
